Renamed to avoid naming clash with parent package. Fix: Field names for edge ordering incorrectly refer to hardcoded values.
